+++ src/evaluate.cpp
@@ -490,7 +490,7 @@ void init_eval(int threads) {
   }
 
   for (Bitboard b = 0ULL; b < 256ULL; b++)
-      BitCount8Bit[b] = count_1s(b);
+      BitCount8Bit[b] = (uint8_t)count_1s(b);
 }
